+++ src/controllers/camera.js
